+++ plugins/ps/js/ps.js
@@ -33,7 +33,7 @@ var ps_show = false;
  */
 function ps_buildTable(xml) {
     var html = "", tree = [], closed = [2];
-    
+
     $("#Plugin_PS #Plugin_PSTable").remove();
 
     html += "  <table id=\"Plugin_PSTable\" class=\"tablemain\" style=\"width:100%;\">\n";
@@ -42,21 +42,23 @@ function ps_buildTable(xml) {
     html += "     <th>" + genlang(3, false, "PS") + "</th>\n";
     html += "     <th style=\"width:80px;\">" + genlang(4, false, "PS") + "</th>\n";
     html += "     <th style=\"width:80px;\">" + genlang(5, false, "PS") + "</th>\n";
-    html += "     <th style=\"width:110px;\">" + genlang(6, false, "PS") + "</th>\n";
+    html += "     <th style=\"width:120px;\">" + genlang(6, false, "PS") + "</th>\n";
+    html += "     <th style=\"width:120px;\">" + genlang(7, false, "PS") + "</th>\n";
     html += "    </tr>\n";
     html += "   </thead>\n";
     html += "   <tbody class=\"tree\">\n";
 
     $("Plugins Plugin_PS Process", xml).each(function ps_getprocess(id) {
-        var close = 0, pid = 0, ppid = 0, name = "", percent = 0, parentId = 0, expanded = 0;
-        name = $(this).attr("Name").replace(/,/g, ",<wbr>").replace(/\s/g, " <wbr>"); /* split long name */
+        var close = 0, pid = 0, ppid = 0, name = "", percent = 0, parentId = 0, expanded = 0, cpu = 0;
+        name = $(this).attr("Name").replace(/,/g, ",<wbr>").replace(/\s/g, " <wbr>").replace(/\./g, ".<wbr>"); /* split long name */
         parentId = parseInt($(this).attr("ParentID"), 10);
         pid = parseInt($(this).attr("PID"), 10);
         ppid = parseInt($(this).attr("PPID"), 10);
         percent = parseInt($(this).attr("MemoryUsage"), 10);
+        cpu = parseInt($(this).attr("CPUUsage"), 10);
         expanded = parseInt($(this).attr("Expanded"), 10);
 
-        html += "    <tr><td><span class=\"treespan\">" + name + "</span></td><td>" + pid + "</td><td>" + ppid + "</td><td>" + createBar(percent) + "</td></tr>\n";
+        html += "    <tr><td><span class=\"treespan\">" + name + "</span></td><td>" + pid + "</td><td>" + ppid + "</td><td>" + createBar(percent) + "</td><td>" + createBar(cpu) + "</td></tr>\n";
         close = tree.push(parentId);
         if (!isNaN(expanded) && (expanded === 0)) {
             closed.push(close);